2, 3, 4

+++ minheap.hh
@@ -8,14 +8,14 @@ using namespace std;
 
 class Heap {
     private:
-        vector< tuple<int, float> > heap;
+        vector< tuple<int, double> > heap;
         int max_relevant;
     public:
         Heap (int mr);
         Heap ();
-        void insert (tuple<int, float> v);
-        tuple<int, float> delmin ();
-        tuple<int, float> peekmin ();
+        void insert (tuple<int, double> v);
+        tuple<int, double> delmin ();
+        tuple<int, double> peekmin ();
         void heapify (int i);
         int test_heap ();
         int size ();
